refactor: Centralize index store opening in IndexStoreService
Coupling analysis will share one IndexStoreDB with LCOM4 (opening
costs seconds), so store existence checks, libIndexStore discovery,
and database creation move out of SemanticLCOMCalculator into a
shared service. Two latent issues are fixed on the way: the database
now waits until fully populated before the first query (a one-shot CLI
has no later chance to catch up), and the database path is keyed to
the store path so different projects can no longer share one database.
