+++ Sources/SwiftComplexityCore/Analysis/SemanticLCOMCalculator.swift
@@ -7,10 +7,6 @@ import SwiftSyntax
 enum LCOMError: LocalizedError {
     case noMembersFound(className: String)
     case parsingFailed(className: String, underlying: Error)
-    case indexStoreNotFound(indexStorePath: String)
-    case libIndexStoreNotFound(searchedPath: String)
-    case toolchainRequired
-    case indexDBInitializationFailed(underlying: Error)
     case symbolNotFound(symbolName: String, usr: String?)
     case queryTimeout(query: String)
 
@@ -20,21 +16,6 @@ enum LCOMError: LocalizedError {
             return "No members found in class '\(className)'"
         case .parsingFailed(let className, let error):
             return "Failed to parse class '\(className)': \(error.localizedDescription)"
-        case .indexStoreNotFound(let indexStorePath):
-            return """
-                Index store not found at '\(indexStorePath)'.
-                Run 'swift build' first to generate the index.
-                """
-        case .libIndexStoreNotFound(let searchedPath):
-            return "libIndexStore not found at: \(searchedPath)"
-        case .toolchainRequired:
-            #if os(Linux)
-                return "--toolchain-path is required for LCOM4 analysis on Linux"
-            #else
-                return "Failed to detect Xcode toolchain. Please specify --toolchain-path"
-            #endif
-        case .indexDBInitializationFailed(let error):
-            return "Failed to initialize IndexStoreDB: \(error.localizedDescription)"
         case .symbolNotFound(let symbolName, let usr):
             let usrInfo = usr.map { " (USR: \($0))" } ?? ""
             return "Symbol '\(symbolName)' not found in index\(usrInfo)"
@@ -104,6 +85,13 @@ actor SemanticLCOMCalculator {
     private let indexStoreDB: IndexStoreDB
     private let indexStorePath: URL
 
+    /// Initialize with an already opened index store, shared with other
+    /// index-backed calculators so the database is populated only once.
+    init(indexStore: SharedIndexStore) {
+        self.indexStoreDB = indexStore.database
+        self.indexStorePath = indexStore.storePath
+    }
+
     /// Initialize with explicit IndexStore path
     /// - Parameters:
     ///   - indexStorePath: Direct path to the IndexStore (e.g., .build/debug/index/store)
@@ -111,20 +99,10 @@ actor SemanticLCOMCalculator {
     ///                    On macOS, if nil, Xcode toolchain is auto-detected.
     ///                    On Linux, this is required.
     init(indexStorePath: URL, toolchainPath: URL? = nil) throws {
-        self.indexStorePath = indexStorePath
-
-        guard FileManager.default.fileExists(atPath: indexStorePath.path) else {
-            throw LCOMError.indexStoreNotFound(indexStorePath: indexStorePath.path)
-        }
-
-        // Get libIndexStore path (platform-specific)
-        let libIndexStorePath = try Self.findLibIndexStore(toolchainPath: toolchainPath)
-
-        self.indexStoreDB = try IndexStoreDB(
-            storePath: indexStorePath.path,
-            databasePath: NSTemporaryDirectory() + "lcom4-index.db",
-            library: IndexStoreLibrary(dylibPath: libIndexStorePath)
-        )
+        let store = try IndexStoreService.open(
+            indexStorePath: indexStorePath, toolchainPath: toolchainPath)
+        self.indexStoreDB = store.database
+        self.indexStorePath = store.storePath
     }
 
     /// Calculates LCOM4 value for Nominal Type (class/struct/actor)
@@ -447,80 +425,6 @@ actor SemanticLCOMCalculator {
         return methodCalls
     }
 
-    // MARK: - Helper Methods
-
-    /// Searches for libIndexStore path
-    /// - Parameter toolchainPath: Optional toolchain path (e.g., ~/.local/share/swiftly/toolchains/swift-6.2).
-    ///   On macOS, if nil, auto-detects from Xcode.
-    ///   On Linux, this is required.
-    /// - Returns: Path to libIndexStore.dylib (macOS) or libIndexStore.so (Linux)
-    private static func findLibIndexStore(toolchainPath: URL?) throws -> String {
-        #if os(Linux)
-            let libName = "libIndexStore.so"
-        #else
-            let libName = "libIndexStore.dylib"
-        #endif
-
-        // If toolchainPath is provided, use it directly
-        // Expected structure: <toolchainPath>/usr/lib/libIndexStore.{so,dylib}
-        if let toolchainPath = toolchainPath {
-            let libPath =
-                toolchainPath
-                .appendingPathComponent("usr")
-                .appendingPathComponent("lib")
-                .appendingPathComponent(libName)
-            guard FileManager.default.fileExists(atPath: libPath.path) else {
-                throw LCOMError.libIndexStoreNotFound(searchedPath: libPath.path)
-            }
-            return libPath.path
-        }
-
-        #if os(macOS)
-            // Auto-detect Xcode toolchain on macOS
-            return try findLibIndexStoreFromXcode()
-        #else
-            // On Linux, toolchainPath is required
-            throw LCOMError.toolchainRequired
-        #endif
-    }
-
-    #if os(macOS)
-        /// Auto-detect libIndexStore from Xcode toolchain (macOS only)
-        private static func findLibIndexStoreFromXcode() throws -> String {
-            // Get Xcode toolchain path using xcrun
-            let process = Process()
-            process.executableURL = URL(fileURLWithPath: "/usr/bin/xcrun")
-            process.arguments = ["--show-sdk-path"]
-
-            let pipe = Pipe()
-            process.standardOutput = pipe
-            process.standardError = pipe
-
-            try process.run()
-            process.waitUntilExit()
-
-            let data = pipe.fileHandleForReading.readDataToEndOfFile()
-            guard
-                let sdkPath = String(data: data, encoding: .utf8)?.trimmingCharacters(
-                    in: .whitespacesAndNewlines)
-            else {
-                throw LCOMError.toolchainRequired
-            }
-
-            // Infer toolchain lib directory from SDK path
-            // /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk
-            // -> /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/libIndexStore.dylib
-            let xcodeAppPath = sdkPath.components(separatedBy: "/Platforms/").first ?? ""
-            let libPath =
-                "\(xcodeAppPath)/Toolchains/XcodeDefault.xctoolchain/usr/lib/libIndexStore.dylib"
-
-            guard FileManager.default.fileExists(atPath: libPath) else {
-                throw LCOMError.libIndexStoreNotFound(searchedPath: libPath)
-            }
-
-            return libPath
-        }
-    #endif
 }
 
 // MARK: - Syntax Visitors (Fallback)
